topologyviewer: add legend image, various bugfixes
This commit removes stations in the topology that are already visible as agents.
A legend image is added.
The traces between interfaces show the mediatype of the link. (Solid for Ethernet, dashed for Wi-Fi, purple dash for station links)
Interfaces are sorted so traces won't visually cross.
A bug with the customdata list is fixed, so clickable objects show the right parameter data in the nodelist.
The resulting plotview is dynamically "scaled" based on the size of the topology.

Signed-off-by: Maarten De Decker <maarten.dedecker@mind.be>