main: network_graph: add blinking AP node logic.
If an AP (Controller or Agent) has stations associated to them, have
them 'blink' by changing between an open-circle and a full-circle on
every other render cycle.

This makes it more obvious to anyone using the topology view which
stations are associated to which APs, especially in a 1-station case.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>