nbapi_rpc.py: replace client_mac arg with empty string
VBSS creation requests do not require a client MAC, but the NBAPI call
expects one anyways (but does not make use of it).

Instead of enforcing a user of this UI to select a STA for a VBSS
creation, for no reason, we instead just pass along a null string to
make the NBAPI function call happy.

Signed-off-by: Tucker Polomik <t.polomik@cablelabs.com>